extract display outside of cmd
