+++ cmd/login.go
@@ -7,8 +7,8 @@ import (
 	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/getsavvyinc/savvy-cli/cmd/browser"
-	"github.com/getsavvyinc/savvy-cli/cmd/display"
 	"github.com/getsavvyinc/savvy-cli/config"
+	"github.com/getsavvyinc/savvy-cli/display"
 	"github.com/getsavvyinc/savvy-cli/savvy_errors"
 	"github.com/spf13/cobra"
 )
